--- a/vllm/engine/arg_utils.py
+++ b/vllm/engine/arg_utils.py
@@ -426,7 +426,8 @@ class EngineArgs:
             self.dtype, self.seed, self.revision, self.code_revision,
             self.tokenizer_revision, self.max_model_len, self.quantization,
             self.quantization_param_path, self.enforce_eager,
-            self.max_context_len_to_capture, self.max_logprobs, self.use_distributed_loading)
+            self.max_context_len_to_capture, self.max_logprobs,
+            self.use_distributed_loading)
         cache_config = CacheConfig(self.block_size,
                                    self.gpu_memory_utilization,
                                    self.swap_space, self.kv_cache_dtype,