@chestnut-Q 2024-03-31
@ywang96 Thank you for your suggestion! I have moved the duplicate codes to utils as per your advice, which resulted in minimal changes to each model file and made them clearer. Additionally, I have added the cli arg `--use-distributed-loading` and expanded support for other models that have a parameter structure similar to llama. If there are any more questions or suggestions, welcome to comment :)