@youkaichao 2024-09-03
close as this becomes stale.

and please see https://github.com/vllm-project/vllm/pull/6127#issuecomment-2221614995 

we recommend using safetensors format, and then we don't need this optimization.